Refactor(YoloSimulation): Invoke Simulation as Subprocess
